funciones modificadoras matriz
