+++ tinkercad/modify.cpp
@@ -1,0 +1,20 @@
+#include <iostream>
+
+using namespace std;
+
+void print_led_matrix(int (*matrix)[8], int size) {
+    for (int i = 0; i < size; i++) {
+        for (int j = 0; j < size; j++) {
+            cout << matrix[i][j] << " ";
+        }
+        cout << endl;
+    }
+}
+
+void fill_matrix(int (*matrix)[8], int size) {
+    for (int i = 0; i < size; i++) {
+        for (int j = 0; j < size; j++) {
+            matrix[i][j] = 0;
+        }
+    }
+}
