+++ tinkercad/modify.h
@@ -1,0 +1,8 @@
+#ifndef MODIFY_H
+#define MODIFY_H
+#include <iostream>
+
+void print_led_matrix(int (*matrix)[8], int size);
+void fill_matrix(int (*matrix)[8], int size);
+
+#endif // MODIFY_H
